NODGrid

FerriteDistributed.NODGrid — Type

NODGrid{dim,C<:AbstractCell,T<:Real} <: AbstractNODGrid{dim}

Scalable non-overlapping distributed grid. This data structure is composed of a local_grid and full topological information on the process boundary, i.e. how vertices, edges and faces are connectedted between processes.

Todo

PartitionedArrays.jl ready constructor via extension

FerriteDistributed.NODGrid — Method

NODGrid(grid_comm::MPI.Comm, grid_to_distribute::Grid{dim,C,T})

Construct a non-overlapping distributed grid from a grid with the SFC induced by the element ordering on a specified MPI communicator. It is assumed that this function is called with exactly the same grid on each MPI process in the communicator.

FerriteDistributed.NODGrid — Method

NODGrid(grid_comm::MPI.Comm, grid_to_distribute::Grid{dim,C,T}, grid_topology::CoverTopology, partitioning::Vector{<:Integer})

Construct a non-overlapping distributed grid from a grid with given topology and partitioning on a specified MPI communicator.

FerriteDistributed.SharedEntity — Type

SharedEntity

Supertype for shared entities.

FerriteDistributed.SharedVertex — Type

SharedVertex <: SharedEntity

A shared vertex induced by a local vertex index and all remote vertex indices on all remote ranks.

FerriteDistributed.SharedFace — Type

SharedFace <: SharedEntity

A shared face induced by a local face index and all remote face indices on all remote ranks.

FerriteDistributed.SharedEdge — Type

SharedEdge <: SharedEntity

A shared edge induced by a local edge index and all remote edge indices on all remote ranks.

FerriteDistributed.remote_entities — Function

remote_entities(::SharedEntity)

Get an iterable of pairs, containing the 1-based rank and a collection of EntityIndices on the rank.

FerriteDistributed.global_comm — Function

global_comm(::AbstractNODGrid)

Get a communicator handle for the full grid.

global_comm(::NODGrid)

Global dense communicator of the distributed grid.

FerriteDistributed.interface_comm — Function

interface_comm(::AbstractNODGrid)

Get a graph comminicator handle for the process boundary.

interface_comm(::NODGrid)

Graph communicator for shared vertices. Guaranteed to be derived from the communicator returned by @global_comm .

FerriteDistributed.global_rank — Function

global_rank(::AbstractNODGrid)

Get the 1-based rank of the global communicator of the grid.

global_rank(::NODGrid)

Get the rank on the global communicator of the distributed grid.

FerriteDistributed.global_nranks — Function

global_nranks(::AbstractNODGrid)

Get the number of ranks of the global communicator of the grid.

global_nranks(::NODGrid)

Get the number of ranks on the global communicator of the distributed grid.