Interface

FerriteDistributed.get_shared_vertices — Function

get_shared_vertices(::AbstractNODGrid)

Get an interable over the shared vertices.

FerriteDistributed.get_shared_vertex — Function

get_shared_vertex(::AbstractNODGrid, ::VertexIndex)
get_shared_vertex(::AbstractNODGrid, ::VertexRepresentation)

Get the shared vertex associated to the vertex, if it exists.

FerriteDistributed.get_shared_edges — Function

getsharededges(::AbstractNODGrid)

Get an interable over the shared edges.

FerriteDistributed.get_shared_edge — Function

get_shared_edge(::AbstractNODGrid, ::EdgeIndex)
get_shared_edge(::AbstractNODGrid, ::EdgeRepresentation)

Get the shared edge associated to the edge, if it exists.

FerriteDistributed.get_shared_faces — Function

get_shared_faces(::AbstractNODGrid)

Get an interable over the shared faces.

FerriteDistributed.get_shared_face — Function

get_shared_face(::AbstractNODGrid, ::FaceIndex)
get_shared_face(::AbstractNODGrid, ::FaceRepresentation)

Get the shared edge associated to the face, if it exists.

FerriteDistributed.is_shared_vertex — Function

is_shared_vertex(::AbstractNODGrid, ::VertexIndex)
is_shared_vertex(::AbstractNODGrid, ::VertexRepresentation)

Check if a VertexIndex is a shared vertex.

FerriteDistributed.is_shared_edge — Function

is_shared_edge(::AbstractNODGrid, ::EdgeIndex)
is_shared_edge(::AbstractNODGrid, ::EdgeRepresentation)

Check if a EdgeIndex is a shared edge.

FerriteDistributed.is_shared_face — Function

is_shared_face(::AbstractNODGrid, ::FaceIndex)
is_shared_face(::AbstractNODGrid, ::FaceRepresentation)

Check if a FaceIndex is a shared face.

FerriteDistributed.getlocalgrid — Function

getlocalgrid(::AbstractNODGrid)

Get the representative local grid containing only a vanilla local grid.