Utils

FerriteDistributed.generate_nod_grid — Function

generate_nod_grid(comm::MPI.Comm, args...)

Helper to directly generate non-overlapping distributed grids, designed to replace the call to generate_grid for use in distributed environments.

FerriteDistributed.create_partitioning — Function

create_partitioning(::Ferrite.AbstractGrid, ::Ferrite.AbstractTopology, nparts::Int, partition_alg)::Vector{Int}

Internal entry point for the construction of Ferrite.jl grid partitionings. Creates a vector where the index corresponds to the cell index of the input grid and the value is the 1-based rank.

FerriteDistributed.compute_owner — Function

compute_owner(::AbstractNODGrid, ::SharedEntity)

Compute which rank (1-based index) owns the corresponding entity.

FerriteDistributed.vtk_shared_edges — Function

vtk_shared_edges(vtk::PVTKGridFile, dgrid::AbstractNODGrid)

Enrich the VTK file with meta information about shared edges.

FerriteDistributed.vtk_shared_faces — Function

vtk_shared_faces(vtk::PVTKGridFile, dgrid::AbstractNODGrid)

Enrich the VTK file with meta information about shared faces.

FerriteDistributed.vtk_shared_vertices — Function

vtk_shared_vertices(vtk::PVTKGridFile, dgrid::AbstractNODGrid)

Enrich the VTK file with meta information about shared vertices.

FerriteDistributed.vtk_partitioning — Function

vtk_partitioning(vtk::PVTKGridFile, dgrid::AbstractNODGrid)

Enrich the VTK file with partitioning meta information.